+++ cmd/graphi/main.go
@@ -93,6 +93,8 @@ func main() {
 		os.Exit(runAgentTool(os.Args[2:], "change-impact"))
 	case "hotspots":
 		os.Exit(runAgentTool(os.Args[2:], "hotspots"))
+	case "search-hybrid":
+		os.Exit(runAgentTool(os.Args[2:], "search-hybrid"))
 	case "related-files":
 		os.Exit(runAgentTool(os.Args[2:], "related-files"))
 	case "change-risk":